fix(PriceService): 정수형 끼리 나누어서 총 수익률이 안나오는 문제 해결

+++ src/main/java/lotto/service/PriceService.java
@@ -32,6 +32,6 @@ public class PriceService {
     }
 
     public float getEarningRate() {
-        return winningPrice.getWinningPrice() / purchasePrice.getPurchasePrice() * PERCENTAGE;
+        return (float) winningPrice.getWinningPrice() / purchasePrice.getPurchasePrice() * PERCENTAGE;
     }
 }
